Add new `embedding` intrinsic
The `embedding` intrinsic takes a statically analyzable collection of values
as its argument, and returns a selection function that selects the most similar
values to some query. Under the hood, the `embedding` intrinsic generates text
embeddings for the `@intent` tags of each value, and performs a vector search
to find similar values.

Rename the `implement` intrinsic back to `generative`.

Remove implicit tool scoping; it's too much of a break from language norms.
Replace it with a configurable `contextTools` compiler import. Add an
async-local `currentTools` to `@toolcog/runtime` that gets used as the
`contextTools` for the REPL.