feat: Configurable inclusion of package and module names in generated IDs
Set `packageId` to a string to override the package component of generated IDs.
Set `packageId` to `false` to exclude the package component of generated IDs.

Set `moduleId` to a string to override the module path component of generated IDs.
Set `moduleId` to `false` to exclude the module path component of generated IDs.