feat: Support more flexible toolkit module definitions.
Default exports from toolkit modules can now be object literals, in addition
to functions. Toolkit functions can now return an immediate object, in addition
to a promise. And a toolkit's `tools` property can now be an immediate array of
tools, a function that returns an array of tools, or a function that returns a
promised array of tools.

Fixed an issue that caused comments to be inadvertently analyzed for the return
types of `defineTool` calls.

Fixed `callSiteToSchema` using an incorrectly spelled `return` property.

All intrinsic transformers now consistently preserve comments that were
attached to the original intrinsic call expression.